+++ jest.setup.js
@@ -1,5 +1,4 @@
-jest.setTimeout(120000); // 2 min
-
+jest.setTimeout(420000); // 7 min
 // add all jest-extended matchers
 const matchers = require("jest-extended");
 expect.extend(matchers);