+++ packages/context/src/ui5-model.ts
@@ -147,10 +147,10 @@ async function createSemanticModelWithFetcher(
           apiJson = await response.json();
           await writeToCache(cacheFilePath, apiJson);
         } else if (response.status === 404) {
-          getLogger().error("Could not find UI5 lib from", { url });
+          getLogger().warn("Could not find UI5 lib from", { url });
           await writeToCache(cacheFilePath, {}); // write dummy file! TODO: how to invalidate?
         } else {
-          getLogger().error("Could not read UI5 lib from", { url });
+          getLogger().warn("Could not read UI5 lib from", { url });
         }
       } else {
         getLogger().info("Reading Cache For UI5 Lib", {
@@ -284,7 +284,7 @@ async function getVersionInfo(
         throw new Error(`Version info request has failed (${url})`);
       }
     } catch (e) {
-      getLogger().error("Could not read version information. " + e, {
+      getLogger().warn("Could not read version information. " + e, {
         url,
       });
     }