+++ test-packages/test-utils/api.d.ts
@@ -92,13 +92,11 @@ export const DEFAULT_UI5_VERSION = "1.71.61";
 
 // TODO: list should be updated continuously!
 export type TestModelVersion =
-  | "1.60.44"
   | /* OOM */ typeof DEFAULT_UI5_VERSION
-  | "1.84.27"
-  | "1.96.11"
-  | "1.105.0"
-  | "1.108.2"
-  | "1.109.1";
+  | "1.84.41"
+  | "1.96.27"
+  | "1.108.26"
+  | "1.114.11";
 
 /**
  * Return a UI5SemanticModel for the specified version.