+++ packages/binding/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ui5-language-assistant/binding
 
+## 1.0.22
+
+### Patch Changes
+
+- a9a2141: Add logger package
+- Updated dependencies [a9a2141]
+  - @ui5-language-assistant/logic-utils@4.0.16
+  - @ui5-language-assistant/context@4.0.22
+  - @ui5-language-assistant/xml-views-completion@4.0.22
+  - @ui5-language-assistant/xml-views-validation@4.0.17
+
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies [910e437]
+  - @ui5-language-assistant/context@4.0.21
+  - @ui5-language-assistant/xml-views-completion@4.0.21
+  - @ui5-language-assistant/xml-views-validation@4.0.16
+  - @ui5-language-assistant/logic-utils@4.0.15
+
 ## 1.0.20
 
 ### Patch Changes